fix: change board input dtype from Int64 to Int32 for TensorRT compatibility
TensorRT has limited Int64 support, causing engine build failures with
"Could not find any implementation for node {ForeignNode[/Cast...]}" errors.
nn.Embedding supports Int32 indices and board values are 0-255, so Int32 is sufficient.

Also adds automatic ONNX Int64→Int32 conversion in build_engine_from_onnx
for backward compatibility with existing ONNX models exported with Int64.

+++ src/maou/app/learning/dl.py
@@ -279,7 +279,7 @@ class Learning:
                 9,
                 9,
             ),
-            dtypes=[torch.int64],
+            dtypes=[torch.int32],
         )
 
         # Compile after freeze + optimizer setup
@@ -295,7 +295,7 @@ class Learning:
                 config.batch_size,
                 9,
                 9,
-                dtype=torch.int64,
+                dtype=torch.int32,
                 device=self.device,
             )
             warmup_compiled_model(self.model, dummy)
